requirements.txt
Django==3.2.17
django-crispy-forms==1.13.0
django-formtools==2.4